+++ src/types.rs
@@ -1,5 +1,0 @@
-//! Type aliases.
-
-use std::sync::{Arc, Mutex};
-
-pub type ArcMutex<T> = Arc<Mutex<T>>;